User Choice of individual or family summaries added
Made it so users can choose between summarising data on and individual-by-individual basis or a family-by-family basis. The thinking is that this way, we can combine siblings to make a better representation of the familial archetype, and look at individual samples if something appears to be off regarding the familial archetype. It may be worthwhile to revisit this to randomly sample an equal number of measurements from each individual being averaged, such that we can reduce the weight one sibling might have relative to another. Also, windorization may prove valuable.